chore(core): rename LLM to Model
