+++ crates/runtime/tests/runtime.rs
@@ -3,7 +3,7 @@
 use compact_str::CompactString;
 use std::collections::BTreeMap;
 use walrus_runtime::{Hook, Runtime, SkillRegistry};
-use wcore::model::{FunctionCall, General, Message, NoopProvider, Tool, ToolCall};
+use wcore::model::{FunctionCall, General, Message, Tool, ToolCall};
 use wcore::{Agent, InMemory, Memory, Skill, SkillTier};
 
 fn echo_tool() -> Tool {
@@ -17,7 +17,7 @@ fn echo_tool() -> Tool {
 
 #[test]
 fn resolve_returns_registered_tools() {
-    let mut rt = Runtime::<()>::new(General::default(), NoopProvider, InMemory::new());
+    let mut rt = Runtime::<()>::new(General::default(), (), InMemory::new());
     rt.register(echo_tool(), |args| async move { args });
     let tools = rt.resolve(&["echo".into()]);
     assert_eq!(tools.len(), 1);
@@ -26,14 +26,14 @@ fn resolve_returns_registered_tools() {
 
 #[test]
 fn resolve_skips_unknown() {
-    let rt = Runtime::<()>::new(General::default(), NoopProvider, InMemory::new());
+    let rt = Runtime::<()>::new(General::default(), (), InMemory::new());
     let tools = rt.resolve(&["missing".into()]);
     assert!(tools.is_empty());
 }
 
 #[tokio::test]
 async fn dispatch_calls_handler() {
-    let mut rt = Runtime::<()>::new(General::default(), NoopProvider, InMemory::new());
+    let mut rt = Runtime::<()>::new(General::default(), (), InMemory::new());
     rt.register(echo_tool(), |args| async move { format!("got: {args}") });
 
     let calls = vec![ToolCall {
@@ -54,7 +54,7 @@ async fn dispatch_calls_handler() {
 
 #[tokio::test]
 async fn dispatch_unknown_tool() {
-    let rt = Runtime::<()>::new(General::default(), NoopProvider, InMemory::new());
+    let rt = Runtime::<()>::new(General::default(), (), InMemory::new());
     let calls = vec![ToolCall {
         id: "call_1".into(),
         index: 0,
@@ -71,20 +71,20 @@ async fn dispatch_unknown_tool() {
 
 #[test]
 fn context_limit_default() {
-    let rt = Runtime::<()>::new(General::default(), NoopProvider, InMemory::new());
-    // NoopProvider's Registry impl always returns 64_000.
+    let rt = Runtime::<()>::new(General::default(), (), InMemory::new());
+    // ()'s Registry impl always returns 64_000.
     assert_eq!(rt.context_limit("any-model"), 64_000);
 }
 
 #[test]
 fn runtime_with_inmemory() {
-    let rt = Runtime::<()>::new(General::default(), NoopProvider, InMemory::new());
+    let rt = Runtime::<()>::new(General::default(), (), InMemory::new());
     assert!(rt.memory().entries().is_empty());
 }
 
 #[test]
 fn remember_tool_registered() {
-    let rt = Runtime::<()>::new(General::default(), NoopProvider, InMemory::new());
+    let rt = Runtime::<()>::new(General::default(), (), InMemory::new());
     let tools = rt.resolve(&["remember".into()]);
     assert_eq!(tools.len(), 1);
     assert_eq!(tools[0].name, "remember");
@@ -92,7 +92,7 @@ fn remember_tool_registered() {
 
 #[tokio::test]
 async fn remember_tool_stores_value() {
-    let rt = Runtime::<()>::new(General::default(), NoopProvider, InMemory::new());
+    let rt = Runtime::<()>::new(General::default(), (), InMemory::new());
 
     let calls = vec![ToolCall {
         id: "call_1".into(),
@@ -116,7 +116,7 @@ async fn system_prompt_includes_memory() {
     let memory = InMemory::new();
     memory.set("user", "Prefers short answers.");
 
-    let mut rt = Runtime::<()>::new(General::default(), NoopProvider, memory);
+    let mut rt = Runtime::<()>::new(General::default(), (), memory);
     rt.add_agent(Agent::new("test").system_prompt("You are helpful."));
 
     // api_messages is private, so test via memory content.
@@ -145,7 +145,7 @@ fn make_test_skill(name: &str, tags: &str, body: &str) -> Skill {
 
 #[test]
 fn runtime_without_skills_unchanged() {
-    let rt = Runtime::<()>::new(General::default(), NoopProvider, InMemory::new());
+    let rt = Runtime::<()>::new(General::default(), (), InMemory::new());
     // No skills set — resolve still works for exact matches.
     let tools = rt.resolve(&["remember".into()]);
     assert_eq!(tools.len(), 1);
@@ -153,7 +153,7 @@ fn runtime_without_skills_unchanged() {
 
 #[test]
 fn resolve_glob_prefix() {
-    let mut rt = Runtime::<()>::new(General::default(), NoopProvider, InMemory::new());
+    let mut rt = Runtime::<()>::new(General::default(), (), InMemory::new());
     rt.register(
         Tool {
             name: "foo_a".into(),
@@ -189,7 +189,7 @@ fn resolve_glob_prefix() {
 
 #[test]
 fn resolve_exact_unchanged() {
-    let mut rt = Runtime::<()>::new(General::default(), NoopProvider, InMemory::new());
+    let mut rt = Runtime::<()>::new(General::default(), (), InMemory::new());
     rt.register(echo_tool(), |args| async move { args });
     let tools = rt.resolve(&["echo".into()]);
     assert_eq!(tools.len(), 1);
@@ -204,8 +204,7 @@ fn skill_body_injected() {
         SkillTier::Bundled,
     );
 
-    let mut rt =
-        Runtime::<()>::new(General::default(), NoopProvider, InMemory::new()).with_skills(registry);
+    let mut rt = Runtime::<()>::new(General::default(), (), InMemory::new()).with_skills(registry);
     rt.add_agent(
         Agent::new("dev")
             .system_prompt("Base prompt.")
@@ -219,7 +218,7 @@ fn skill_body_injected() {
 #[test]
 fn skill_tools_registered() {
     // Verify that skills' allowed_tools are available via resolve.
-    let mut rt = Runtime::<()>::new(General::default(), NoopProvider, InMemory::new());
+    let mut rt = Runtime::<()>::new(General::default(), (), InMemory::new());
     rt.register(echo_tool(), |args| async move { args });
 
     // The skill lists "echo" in allowed_tools — it should resolve.
@@ -247,7 +246,7 @@ fn hook_default_flush_prompt() {
 
 #[test]
 fn set_skills_on_existing_runtime() {
-    let mut rt = Runtime::<()>::new(General::default(), NoopProvider, InMemory::new());
+    let mut rt = Runtime::<()>::new(General::default(), (), InMemory::new());
     let registry = SkillRegistry::new();
     rt.set_skills(registry);
     // Resolve still works after setting skills.
@@ -259,7 +258,7 @@ fn set_skills_on_existing_runtime() {
 
 #[tokio::test]
 async fn send_to_unknown_agent_fails() {
-    let mut rt = Runtime::<()>::new(General::default(), NoopProvider, InMemory::new());
+    let mut rt = Runtime::<()>::new(General::default(), (), InMemory::new());
     let result = rt.send_to("unknown", Message::user("hello")).await;
     assert!(result.is_err());
     assert!(result.unwrap_err().to_string().contains("not registered"));
@@ -267,7 +266,7 @@ async fn send_to_unknown_agent_fails() {
 
 #[test]
 fn clear_session_removes() {
-    let mut rt = Runtime::<()>::new(General::default(), NoopProvider, InMemory::new());
+    let mut rt = Runtime::<()>::new(General::default(), (), InMemory::new());
     rt.add_agent(Agent::new("test").system_prompt("hello"));
     // clear_session on a non-existent session is a no-op.
     rt.clear_session("test");
@@ -276,7 +275,7 @@ fn clear_session_removes() {
 
 #[test]
 fn resolve_tools_returns_pairs() {
-    let mut rt = Runtime::<()>::new(General::default(), NoopProvider, InMemory::new());
+    let mut rt = Runtime::<()>::new(General::default(), (), InMemory::new());
     rt.register(echo_tool(), |args| async move { args });
     let resolved = rt.resolve_tools(&["echo".into()]);
     assert_eq!(resolved.len(), 1);
@@ -285,7 +284,7 @@ fn resolve_tools_returns_pairs() {
 
 #[test]
 fn resolve_tools_glob() {
-    let mut rt = Runtime::<()>::new(General::default(), NoopProvider, InMemory::new());
+    let mut rt = Runtime::<()>::new(General::default(), (), InMemory::new());
     rt.register(
         Tool {
             name: "foo_a".into(),
@@ -312,7 +311,7 @@ fn resolve_tools_glob() {
 
 #[test]
 fn resolve_returns_schemas_only() {
-    let mut rt = Runtime::<()>::new(General::default(), NoopProvider, InMemory::new());
+    let mut rt = Runtime::<()>::new(General::default(), (), InMemory::new());
     rt.register(echo_tool(), |args| async move { args });
     let tools = rt.resolve(&["echo".into()]);
     assert_eq!(tools.len(), 1);
