+++ crates/runtime/tests/team.rs
@@ -1,7 +1,7 @@
 //! Tests for team composition.
 
 use walrus_runtime::{Runtime, build_team, extract_input, worker_tool};
-use wcore::model::{General, NoopProvider};
+use wcore::model::General;
 use wcore::{Agent, InMemory};
 
 #[test]
@@ -33,7 +33,7 @@ fn worker_tool_builds_tool() {
 
 #[test]
 fn build_team_registers_workers_as_tools() {
-    let mut rt = Runtime::<()>::new(General::default(), NoopProvider, InMemory::new());
+    let mut rt = Runtime::<()>::new(General::default(), (), InMemory::new());
     let leader = Agent::new("leader")
         .system_prompt("You coordinate.")
         .description("coordinator");
@@ -60,7 +60,7 @@ fn build_team_registers_workers_as_tools() {
 
 #[test]
 fn build_team_adds_worker_agents() {
-    let mut rt = Runtime::<()>::new(General::default(), NoopProvider, InMemory::new());
+    let mut rt = Runtime::<()>::new(General::default(), (), InMemory::new());
     let leader = Agent::new("leader")
         .system_prompt("You coordinate.")
         .description("coordinator");
@@ -78,7 +78,7 @@ fn build_team_adds_worker_agents() {
 
 #[tokio::test]
 async fn worker_handler_parses_input() {
-    let mut rt = Runtime::<()>::new(General::default(), NoopProvider, InMemory::new());
+    let mut rt = Runtime::<()>::new(General::default(), (), InMemory::new());
     let leader = Agent::new("leader")
         .system_prompt("You coordinate.")
         .description("coordinator");
@@ -106,7 +106,7 @@ async fn worker_handler_parses_input() {
 
 #[test]
 fn build_team_worker_tool_descriptions() {
-    let mut rt = Runtime::<()>::new(General::default(), NoopProvider, InMemory::new());
+    let mut rt = Runtime::<()>::new(General::default(), (), InMemory::new());
     let leader = Agent::new("leader")
         .system_prompt("You coordinate.")
         .description("coordinator");
